Add updated logic to also handle the hardware back button
