+++ src/Store.js
@@ -8,8 +8,8 @@ import { persistStore, persistReducer } from 'redux-persist';
 import { applyMiddleware, createStore } from 'redux';
 import thunk from 'redux-thunk';
 
+import { navigationMiddleware } from './navigation';
 import reducers from './reducers';
-import { RootNavigator } from './navigation';
 
 const persistConfig = {
   keyPrefix: '',
@@ -33,34 +33,6 @@ const persistConfig = {
 };
 
 const persistedReducer = persistReducer(persistConfig, reducers);
-
-/**
- * Simple middleware which intercepts navigation actions and calls a function
- * on the apps root navigator to force navigation.
- *
- * Note: This was added when transitioning from react-navigation v4, where
- * navigation was handled through redux - to v5 where the previous integration
- * method was not possible.
- *
- */
-const navigationMiddleware = () => next => action => {
-  const { type } = action;
-
-  if (type === 'Navigation/NAVIGATE') {
-    const { routeName, params } = action;
-    RootNavigator.navigate(routeName, params);
-  }
-  if (type === 'Navigation/REPLACE') {
-    const { routeName, params } = action;
-    RootNavigator.replace(routeName, params);
-  }
-
-  if (type === 'Navigation/BACK') {
-    RootNavigator.goBack();
-  }
-
-  next(action);
-};
 
 const store = createStore(persistedReducer, {}, applyMiddleware(thunk, navigationMiddleware));
