+++ src/main.js
@@ -4,18 +4,23 @@
  */
 
 import React from 'react';
-import { AppRegistry, YellowBox } from 'react-native';
+import { AppRegistry, YellowBox, BackHandler } from 'react-native';
 import { Provider } from 'react-redux';
 import { NavigationContainer } from '@react-navigation/native';
 import { ErrorHandler } from 'redux-persist-error-handler';
 
 import { name as appName } from '../app.json';
 import { store, persistedStore } from './Store';
-import { rootNavigatorRef } from './navigation';
+import { backHandler, rootNavigatorRef } from './navigation';
 
 import MSupplyMobileApp from './mSupplyMobileApp';
 
-function App() {
+const App = () => {
+  React.useEffect(() => {
+    BackHandler.addEventListener('hardwareBackPress', backHandler(store));
+    return () => BackHandler.removeEventListener('hardwareBackPress', backHandler(store));
+  }, []);
+
   return (
     <ErrorHandler persistedStore={persistedStore}>
       <Provider store={store}>
@@ -25,7 +30,7 @@ function App() {
       </Provider>
     </ErrorHandler>
   );
-}
+};
 
 YellowBox.ignoreWarnings(['Setting a timer']);
