+++ src/widgets/BackButton.js
@@ -4,14 +4,14 @@
  */
 
 import React from 'react';
-import { StyleSheet, Text, TouchableOpacity, BackHandler } from 'react-native';
+import { StyleSheet, Text, TouchableOpacity } from 'react-native';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
 import { debounce } from '../utilities';
-import { UIDatabase } from '../database';
+
 import { getRouteTitle, RootNavigator } from '../navigation';
-import { PrescriptionActions } from '../actions/PrescriptionActions';
+
 import { goBack } from '../navigation/actions';
 
 import { BackIcon } from './icons';
@@ -29,38 +29,22 @@ const mapStateToProps = state => {
 };
 
 const mapDispatchToProps = dispatch => {
-  const deletePrescriptions = () => dispatch(PrescriptionActions.deletePrescription());
-  const onBack = () => dispatch(goBack());
+  const onBack = debounce(() => dispatch(goBack()), 500, true);
 
-  return { deletePrescriptions, onBack };
+  return { onBack };
 };
 
-export const BackButtonComponent = ({ title, deletePrescriptions, onBack }) => {
-  const handleBackEvent = () => {
-    UIDatabase.write(() => {
-      const prescriptions = UIDatabase.objects('Prescription').filtered('status != "finalised"');
-      UIDatabase.delete('Transaction', prescriptions);
-      deletePrescriptions();
-    });
-
-    if (!RootNavigator.canGoBack()) BackHandler.exitApp();
-    else onBack();
-  };
-
-  const debouncedBackEventHandler = React.useCallback(debounce(handleBackEvent, 5000, true), []);
-
-  return (
-    <TouchableOpacity onPress={debouncedBackEventHandler}>
+export const BackButtonComponent = ({ title, onBack }) =>
+  RootNavigator.canGoBack() && (
+    <TouchableOpacity onPress={onBack}>
       <FlexRow flex={1} justifyContent="center" alignItems="center">
-        {RootNavigator.canGoBack() && <BackIcon />}
+        <BackIcon />
         <Text style={localStyles.title}>{title}</Text>
       </FlexRow>
     </TouchableOpacity>
   );
-};
 
 BackButtonComponent.propTypes = {
-  deletePrescriptions: PropTypes.func.isRequired,
   onBack: PropTypes.func.isRequired,
   title: PropTypes.string.isRequired,
 };
